Removed obsolete count-matrix code and tests
