--- msmtools/estimation/sparse/testfiles/C_13_lag.dat
@@ -1,20 +1,0 @@
-2 0 0 1 2 4 1 1 2 4 2 2 3 1 3 2 2 2 2 4
-5 4 3 4 3 3 1 2 1 3 2 1 1 5 1 1 3 1 1 1
-2 2 3 2 1 2 0 4 3 2 1 2 2 1 2 1 1 3 1 2
-1 6 1 3 4 4 1 3 1 3 1 1 1 3 1 0 2 1 0 4
-3 1 1 1 1 2 1 0 1 4 2 2 2 3 5 3 2 1 0 1
-2 2 2 2 3 1 0 2 1 1 1 1 2 3 2 1 2 4 3 5
-4 0 1 2 1 1 2 2 1 3 2 2 1 3 3 0 1 0 2 0
-1 1 1 3 1 0 3 4 3 7 2 1 2 1 1 3 1 2 2 2
-3 2 0 3 1 0 0 1 2 0 1 1 3 2 4 1 3 0 1 5
-0 2 2 4 0 2 4 2 3 0 4 2 4 3 2 3 0 3 3 1
-2 3 2 2 3 1 3 2 1 3 3 2 2 2 2 1 1 2 3 1
-1 0 5 2 2 2 0 3 0 2 2 4 1 2 1 2 3 2 3 1
-0 5 2 3 2 3 1 1 3 1 1 1 2 1 2 1 0 3 0 6
-4 5 4 0 3 1 4 0 1 0 4 1 0 3 4 0 2 5 4 0
-5 1 1 2 0 2 1 2 0 4 0 3 2 2 2 6 3 3 3 0
-0 1 4 1 1 2 2 3 2 1 2 3 3 1 0 0 1 1 0 1
-0 1 1 2 2 0 2 3 0 1 3 4 2 2 0 2 0 1 4 1
-2 4 2 4 2 3 0 3 2 3 3 4 1 2 1 0 1 2 0 1
-1 2 1 0 3 3 3 2 2 1 3 0 1 1 6 1 1 0 2 2
-2 4 1 0 1 4 2 1 4 1 2 1 3 3 1 1 2 4 1 3
