--- msmtools/estimation/sparse/testfiles/C_13_sliding.dat
@@ -1,20 +1,0 @@
-22 32 20 31 25 27 22 22 24 27 35 28 22 27 32 29 19 22 24 33
-34 33 19 42 26 21 11 21 26 31 30 28 25 28 20 15 28 24 20 18
-32 26 28 34 28 25 15 23 27 22 29 25 29 21 25 20 19 31 24 25
-19 32 33 28 31 23 22 39 25 37 21 28 28 19 15 24 21 27 26 28
-29 22 18 15 23 28 28 23 17 27 28 32 38 31 36 26 26 20 20 24
-33 27 26 17 29 16 25 18 23 22 22 21 30 20 29 27 34 30 18 24
-29 20 37 25 25 18 30 27 27 23 20 22 26 28 26 25 21 23 28 22
-19 25 28 18 23 22 34 25 28 28 23 26 22 17 23 28 21 16 33 23
-22 25 22 28 32 29 30 23 32 9 20 20 26 30 36 29 31 22 27 23
-30 25 21 34 13 26 34 28 21 27 26 27 22 16 27 28 25 29 24 24
-20 30 25 22 26 27 27 29 23 20 26 28 19 22 16 19 28 20 34 31
-22 22 37 25 27 28 22 28 25 25 24 26 13 25 25 34 21 37 25 20
-28 23 24 33 33 30 20 22 23 29 19 24 28 25 25 24 17 26 20 25
-25 21 26 19 19 23 22 23 30 26 24 26 24 28 32 27 19 26 24 21
-28 18 25 21 30 21 34 30 27 22 22 27 26 19 28 31 22 22 29 21
-20 21 31 32 25 20 27 24 27 23 37 24 25 30 16 24 29 26 17 21
-26 23 24 22 20 21 23 22 29 25 21 23 29 19 27 24 25 18 23 25
-31 24 20 27 22 29 31 17 17 42 26 27 21 25 20 24 18 26 22 21
-25 30 20 29 22 21 25 23 28 21 19 25 22 21 27 22 21 24 28 35
-30 23 24 24 32 36 20 13 38 20 20 23 23 33 18 20 26 20 22 21
