--- msmtools/estimation/sparse/testfiles/C_1_lag.dat
@@ -1,20 +1,0 @@
-35 27 28 28 34 33 19 25 19 27 26 25 28 22 27 26 20 25 32 20
-32 18 27 26 21 25 21 26 41 35 16 18 26 24 26 21 26 25 24 24
-24 31 21 36 18 28 31 24 29 28 25 21 22 22 23 26 21 25 24 29
-27 27 25 29 37 22 30 29 36 28 24 29 29 17 29 18 26 27 18 19
-34 25 24 22 27 25 27 16 24 23 18 37 34 28 21 31 23 20 29 23
-23 25 30 33 32 16 23 17 26 21 19 33 22 30 23 22 20 31 18 28
-20 21 23 27 24 20 24 18 30 29 24 24 25 27 35 22 24 27 28 30
-32 26 19 22 20 23 23 27 16 32 28 26 21 31 14 26 32 23 21 20
-31 23 29 24 40 26 32 23 29 22 31 32 23 18 20 27 21 20 27 19
-27 19 30 24 25 22 20 27 29 25 23 28 21 25 29 28 28 25 27 25
-28 27 26 25 19 23 25 22 23 23 20 32 17 30 28 30 22 29 24 19
-28 31 20 27 30 32 27 18 21 22 21 23 29 30 41 25 19 18 24 25
-22 28 27 23 19 28 19 27 18 25 28 31 25 31 20 25 23 17 28 34
-28 18 27 34 22 18 28 28 23 22 31 14 23 22 30 24 16 25 26 26
-20 27 19 21 24 28 35 28 22 30 25 24 30 16 25 27 23 26 24 30
-19 20 32 23 27 29 27 30 27 29 25 29 18 30 14 22 26 28 22 23
-23 22 12 27 24 30 27 28 25 29 27 17 21 19 24 21 29 25 22 19
-25 31 31 21 16 22 30 22 29 26 26 24 26 13 32 21 28 28 20 19
-25 29 32 22 23 29 19 27 23 17 20 27 29 25 18 31 17 23 23 30
-23 27 26 32 29 13 16 20 27 14 35 17 29 24 25 27 27 23 28 24
