--- msmtools/estimation/sparse/testfiles/C_7_lag.dat
@@ -1,20 +1,0 @@
-2 5 2 3 4 2 3 6 3 4 5 2 4 7 4 4 7 2 2 4
-7 2 4 1 4 3 3 3 7 4 3 5 4 6 4 5 1 4 4 3
-4 2 1 3 1 6 5 3 2 4 7 6 3 3 3 5 2 1 4 1
-2 5 4 4 7 5 5 1 4 4 1 3 1 0 4 7 2 5 0 1
-2 4 2 5 3 8 4 3 5 4 7 2 2 2 4 5 3 5 4 1
-3 2 3 9 5 3 4 0 3 4 1 3 4 3 6 3 9 7 5 5
-3 2 4 4 5 1 3 5 4 3 3 3 5 5 5 4 0 1 2 7
-2 6 3 1 7 1 2 1 3 6 6 2 3 2 1 5 1 4 1 2
-3 3 3 2 4 5 3 5 2 3 3 9 3 3 3 4 2 8 5 2
-5 5 1 7 2 5 5 1 4 3 7 3 2 6 5 3 3 4 1 8
-9 2 6 1 4 1 1 2 5 4 2 6 1 5 7 5 3 4 5 0
-3 0 4 7 4 9 2 3 2 7 3 6 3 4 1 2 5 1 7 2
-2 5 4 1 4 3 3 1 4 9 2 3 1 1 6 1 1 2 4 3
-2 3 3 4 3 5 4 3 6 5 2 2 3 4 4 2 0 3 6 3
-4 6 3 5 4 4 6 4 6 4 4 1 6 0 8 3 4 5 6 4
-6 6 6 1 1 3 3 5 0 6 4 4 2 4 5 2 2 4 5 5
-6 5 2 4 1 3 4 3 4 2 3 2 3 2 2 1 1 4 4 2
-5 6 4 1 4 4 3 4 6 1 3 6 5 3 3 5 4 2 3 4
-3 2 5 0 6 6 4 6 2 1 5 3 2 4 5 7 2 5 3 3
-2 6 2 2 2 6 2 0 3 2 2 4 3 2 7 1 6 5 3 1
