--- msmtools/estimation/sparse/testfiles/C_7_sliding.dat
@@ -1,20 +1,0 @@
-23 31 17 31 34 22 34 27 19 29 18 34 29 22 23 26 21 23 28 34
-22 21 26 27 24 20 27 20 30 24 22 32 30 30 25 17 35 23 24 22
-25 20 30 17 29 26 23 31 26 31 33 21 31 24 24 25 26 21 21 24
-33 32 22 23 36 16 31 32 40 17 23 31 27 18 20 22 20 35 23 25
-20 31 25 26 29 34 28 23 31 25 26 26 30 21 27 25 19 26 21 18
-19 21 25 33 29 20 19 22 20 23 32 30 21 24 28 23 37 22 23 20
-22 21 24 27 24 14 27 33 24 19 30 25 27 21 22 28 25 32 29 28
-38 24 28 23 24 23 32 18 30 26 22 19 23 24 15 25 22 23 18 25
-27 20 34 25 29 20 26 29 24 35 25 30 29 24 23 28 17 23 25 24
-19 35 16 40 16 33 27 16 25 23 24 36 27 19 26 25 21 29 18 32
-29 22 34 22 18 23 16 17 27 16 23 24 23 34 33 34 22 22 33 20
-29 22 25 29 24 32 22 25 24 26 17 24 18 35 23 24 40 24 27 21
-22 28 29 22 21 31 27 23 26 28 25 27 12 22 21 33 21 16 33 31
-25 21 21 38 29 28 20 21 23 30 20 18 22 29 26 24 13 26 27 24
-33 27 28 27 26 20 28 22 22 33 26 23 24 19 31 20 24 20 24 26
-30 27 37 28 19 23 16 34 24 26 17 20 28 30 28 20 21 24 28 20
-28 35 24 25 21 20 24 13 31 27 26 22 27 19 22 20 15 19 24 27
-23 21 20 19 29 34 29 26 27 21 33 25 19 20 20 23 20 24 26 31
-35 24 22 20 23 21 25 30 23 20 27 21 25 26 26 32 22 32 16 19
-23 19 21 24 27 32 21 19 21 28 23 23 26 23 40 26 30 25 21 14
